customize(router): init Websocket before auto-signin
Initialize the Websocket before the login.

Refactor Websocket to own module to use it in the auth module.

== Reproducer

1. Logout/de-auth by deleting token in Cookies and localStorage
2. Go to /auth
3. Enter credentials and Sign in
4. Get a "login successful" toast

Actual: you remain at /auth
Expected: route to /

== Background

I forgot to initialize the Websocket after successful authentication in

   ac37ed2711d4e5c4b874a5ff35c6535880b4457e
   customize(router): move Websocket setup after session validation
